restructured kafka simple example
