+++ src/main/java/com/ibm/hello/service/kafka/simple/ConsumerCreator.java
@@ -1,4 +1,4 @@
-package com.ibm.hello.service;
+package com.ibm.hello.service.kafka.simple;
 
 import java.util.Collections;
 import java.util.Properties;
@@ -8,7 +8,8 @@ import org.apache.kafka.clients.consumer.ConsumerConfig;
 import org.apache.kafka.clients.consumer.KafkaConsumer;
 import org.apache.kafka.common.serialization.LongDeserializer;
 import org.apache.kafka.common.serialization.StringDeserializer;
-import com.ibm.hello.config.SimpleKafkaConstants;
+
+import com.ibm.hello.config.kafka.simple.SimpleKafkaConstants;
 
 
 public class ConsumerCreator {
